Split up wl_shm & wl_compositor, better interface
